packages/happy-voice-agent/requirements.txt
livekit-agents[codecs,silero,turn-detector]~=1.2.0
python-dotenv~=1.0
pytest~=8.3
